requirements.txt
streamlit
pandas
numpy
sentence-transformers
scikit-learn
pymupdf
pyarrow
requests
faiss-cpu
